add config package and message templates
