+++ sdk/src/unplug/config/__init__.py
@@ -1,13 +1,16 @@
-"""Re-export config from unplug.config (backward compatibility)."""
+"""Configuration models and loading."""
 
 from __future__ import annotations
 
 from unplug.config.guard import GuardConfig, PipelineConfig, ScannerConfig, ThresholdConfig
+from unplug.config.limits import LimitConfig, LimitViolation
 from unplug.config.loader import build_config, load, load_from_env, load_from_file
 from unplug.config.messages import MessageConfig
 
 __all__ = [
     "GuardConfig",
+    "LimitConfig",
+    "LimitViolation",
     "MessageConfig",
     "PipelineConfig",
     "ScannerConfig",
